Fix calls to get_es_metadata in checks/audit_checks.py to work when ES_HOST_LOCAL is set.
