+++ pyproject.toml
@@ -12,7 +12,7 @@ packages = [
 python = ">=3.8,<3.12"
 cron-descriptor = "^1.4.0"
 click = "^7.1.2"
-dcicutils = "^8.0.0"
+dcicutils = "^8.8.4"
 PyJWT = "^2.5.0"
 Jinja2 = "^3.1.2"
 MarkupSafe = "^2.1.3"
@@ -27,7 +27,7 @@ pytz = "^2020.1"
 tibanna-ff = "^3.2.0"
 ## adding foursight-core
 # use below for deployment
-foursight-core = "^5.4.1"
+foursight-core = "^5.4.2"
 # use below for tests but not for deployment
 # foursight-core = { git = "https://github.com/4dn-dcic/foursight-core.git", branch="master" }
 # Need pytest-redis 3.0.2 or higher for pytest 7.4.2 (or higher).
